requirements.txt
colorama==0.3.6
Django==1.9.2
djangorestframework==3.3.2
flake8==2.5.4
funcsigs==0.4
mccabe==0.4.0
mock==1.3.0
oauthlib==1.0.3
pbr==1.8.1
pep8==1.7.0
py==1.4.31
pyflakes==1.0.0
PyJWT==1.4.0
pytest==2.8.7
pytest-django==2.9.1
six==1.10.0
wheel==0.24.0
